Fixed 2 bugs (both severe) 1) BlockMapper did not correctly handle longs resulting in an overflowed Int with a negative file position as a result 2) CBUtil separate method returned the smallest value instead of the largest value when no separator could be found. This effected all types except strings/char arrays

+++ src/main/java/org/mellowtech/core/io/BlockMapper.java
@@ -45,7 +45,7 @@ public class BlockMapper {
   }
 
   public void expand(int toIndex) throws IOException {
-    int pos = getBufferPos(toIndex);
+    long pos = getBufferPos(toIndex);
     for (int i = blocks.size(); i <= pos; i++) {
       long filePos = offset + (pos * blockSize * blocksToMap);
       blocks.add(fc.map(FileChannel.MapMode.READ_WRITE, filePos, blockSize * blocksToMap));
